+++ CustomInputStream/ZInputStream.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Runtime.InteropServices;
 
-using MonoTouch.CoreFoundation;
-using MonoTouch.Foundation;
+using CoreFoundation;
+using Foundation;
 
 namespace InputStreamTest
 {
@@ -40,8 +40,7 @@ namespace InputStreamTest
 		{
 			return total_length > read_length;
 		}
-	
-		public override int Read (IntPtr buffer, uint len) 
+		public override nint Read (IntPtr buffer, nuint len)
 		{
 			int actual = Math.Min ((int) len, (int) (total_length - read_length));
 			
@@ -59,7 +58,7 @@ namespace InputStreamTest
 			return actual;
 		}
 		
-		protected unsafe override bool GetBuffer (out IntPtr buffer, out uint len)
+		protected unsafe override bool GetBuffer (out IntPtr buffer, out nuint len)
 		{
 			// Just call the base implemention (which will return false)
 			return base.GetBuffer (out buffer, out len);
